JwtUser class refactored, JWT config moved to separate JwtTokenConfig class, added Auth URL to readme, warnings foxed
