+++ src/main/java/com/pet/bankservice/security/JwtUserDetailsService.java
@@ -2,7 +2,6 @@ package com.pet.bankservice.security;
 
 import com.pet.bankservice.entity.User;
 import com.pet.bankservice.security.jwt.JwtUser;
-import com.pet.bankservice.security.jwt.JwtUserMapper;
 import com.pet.bankservice.service.UserService;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -21,7 +20,7 @@ public class JwtUserDetailsService implements UserDetailsService {
     @Override
     public JwtUser loadUserByUsername(String phoneNumber) throws UsernameNotFoundException {
         User user = userService.findByPhoneNumber(phoneNumber);
-        JwtUser jwtUser = JwtUserMapper.create(user);
+        JwtUser jwtUser = JwtUser.create(user);
         log.info("IN UserDetailsService: user found by phonenumber={}", phoneNumber);
         return jwtUser;
     }
